+++ CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/basaltinc/knapsack/compare/v1.1.1...v1.1.2) (2019-04-19)
+
+**Note:** Version bump only for package @knapsack/monorepo
+
+
+
+
+
+## [1.1.1](https://github.com/basaltinc/knapsack/compare/v1.1.0...v1.1.1) (2019-04-19)
+
+**Note:** Version bump only for package @knapsack/monorepo
+
+
+
+
+
+# [1.1.0](https://github.com/basaltinc/knapsack/compare/v1.0.9...v1.1.0) (2019-04-18)
+
+
+### Features
+
+* added beginning ui settings page ([d999cca](https://github.com/basaltinc/knapsack/commit/d999cca))
+* complete ui settings page ([136e327](https://github.com/basaltinc/knapsack/commit/136e327))
+
+
+
+
+
 ## [1.0.9](https://github.com/basaltinc/knapsack/compare/v1.0.8...v1.0.9) (2019-04-18)
 
 **Note:** Version bump only for package @knapsack/monorepo