+++ components/api-demo/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/basaltinc/knapsack/compare/v1.1.1...v1.1.2) (2019-04-19)
+
+**Note:** Version bump only for package @knapsack/api-demo
+
+
+
+
+
 # [1.0.0](https://github.com/basaltinc/bedrock/compare/v0.45.3...v1.0.0) (2019-04-16)