+++ knapsack/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/basaltinc/knapsack/compare/v1.1.1...v1.1.2) (2019-04-19)
+
+**Note:** Version bump only for package @basalt/knapsack
+
+
+
+
+
 ## [1.0.7](https://github.com/basaltinc/knapsack/compare/v1.0.6...v1.0.7) (2019-04-17)